Fixed import issue
This is the version that was uploaded to the MegaJam yesterday. I ran into a problem trying to upload the Project Nature spruce as the textures exceeded the GitHub file limit, so these have been omitted. This is all okay, however, as I'm going to switch to some improved foliage with the next iteration in any case. :smile: